+++ src/components/cascade-picker-view/index.ts
@@ -1,0 +1,5 @@
+import { CascadePickerView } from './cascade-picker-view'
+
+export type { CascadePickerViewProps } from './cascade-picker-view'
+
+export default CascadePickerView
