+++ src/components/cascade-picker/demos/index.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
-import { DemoBlock, sleep } from 'demos'
+import { DemoBlock } from 'demos'
 import { Button, CascadePicker } from 'antd-mobile'
-import { CascadePickerOption } from '../../cascade-picker'
+import { AsyncDemo } from './async-demo'
 
 function CascadePickerDemo() {
   const [visible, setVisible] = useState(false)
@@ -56,62 +56,6 @@ function CascadePickerDemo() {
         }}
         onSelect={val => {
           console.log('onSelect', val)
-        }}
-      />
-    </>
-  )
-}
-
-function AsyncDemo() {
-  const [visible, setVisible] = useState(false)
-  async function mockApi(key: string) {
-    await sleep(1000)
-    return [key + '1', key + '2', key + '3']
-  }
-  const [options, setOptions] = useState<CascadePickerOption[]>([
-    {
-      label: 'A',
-      value: 'A',
-      children: [],
-    },
-    {
-      label: 'B',
-      value: 'B',
-      children: [],
-    },
-  ])
-
-  return (
-    <>
-      <Button
-        onClick={() => {
-          setVisible(true)
-        }}
-      >
-        选择
-      </Button>
-      <CascadePicker
-        options={options}
-        visible={visible}
-        onClose={() => {
-          setVisible(false)
-        }}
-        onSelect={val => {
-          const key = val[0]
-          if (!key) return
-          mockApi(key).then(data => {
-            setOptions(options =>
-              options.map(option => {
-                if (option.value === key) {
-                  return {
-                    ...option,
-                    children: data.map(x => ({ label: x, value: x })),
-                  }
-                }
-                return option
-              })
-            )
-          })
         }}
       />
     </>
