+++ src/components/date-picker-view/index.ts
@@ -1,3 +1,5 @@
 import { DatePickerView } from './date-picker-view'
 
+export type { DatePickerViewProps } from './date-picker-view'
+
 export default DatePickerView
