+++ src/index.ts
@@ -8,6 +8,7 @@ export { default as Badge } from './components/badge'
 export { default as Button } from './components/button'
 export { default as Card } from './components/card'
 export { default as CascadePicker } from './components/cascade-picker'
+export { default as CascadePickerView } from './components/cascade-picker-view'
 export { default as CheckList } from './components/check-list'
 export { default as Checkbox } from './components/checkbox'
 export { default as Collapse } from './components/collapse'
